pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gutzkow-digitale-edition"
description = ''
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "Mark Hall", email = "mark.hall@work.room3b.eu" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "jupyter-book>=0.15.1",
  "lxml>=4.9.2,<5.0.0",
  "sphinx-favicon>=1.0.0,<2.0.0",
  "sphinx-design<1.0.0",
]
dynamic = [
  "version"
]

[project.urls]
Documentation = "https://github.com/unknown/gutzkow-digitale-edition#readme"
Issues = "https://github.com/unknown/gutzkow-digitale-edition/issues"
Source = "https://github.com/unknown/gutzkow-digitale-edition"

[tool.hatch.version]
path = "gutzkow_digitale_edition/__about__.py"

[tool.hatch.envs.default.scripts]
build = "jb build content"

[tool.hatch.envs.dev]
extra-dependencies = [
  "livereload",
]

[tool.hatch.envs.dev.scripts]
server = "python dev-server.py"
